# src/M'baara-openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M'baara API",
    "description": "API documentation for the **M'baara** application.\n\n**Authentication:** Most endpoints require the `api_key` header. If using Swagger UI, click the **Authorize** button at the top of the page and enter your API key before trying any endpoint.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local API server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/entities/Language": {
      "get": {
        "tags": [
          "Entities - Language"
        ],
        "summary": "List Language records",
        "operationId": "list_Language",
        "description": "Retrieve a list of Language records with optional filtering, sorting, and pagination.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON query filter, e.g. {\"status\":\"active\"}"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip (pagination)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Field name to sort by. Prefix with '-' for descending order, e.g. -created_date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities - Language"
        ],
        "summary": "Create a Language record",
        "operationId": "create_Language",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              },
              "example": {
                "code": "Example code",
                "name": "Example name",
                "name_fr": "Example name_fr"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Language"
        ],
        "summary": "Delete multiple Language records",
        "operationId": "delete_many_Language",
        "description": "**Warning:** Passing an empty `{}` query will delete ALL records in this entity.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Query filter to match records to delete"
              },
              "example": {
                "code": "Example code"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deleted": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/Language/{Language_id}": {
      "get": {
        "tags": [
          "Entities - Language"
        ],
        "summary": "Get a Language record by ID",
        "operationId": "get_Language",
        "parameters": [
          {
            "name": "Language_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities - Language"
        ],
        "summary": "Update a Language record",
        "operationId": "update_Language",
        "parameters": [
          {
            "name": "Language_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Language"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Language"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Language"
        ],
        "summary": "Delete a Language record",
        "operationId": "delete_Language",
        "description": "Soft-deletes the record. Use the restore endpoint to undo.",
        "parameters": [
          {
            "name": "Language_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/Contribution": {
      "get": {
        "tags": [
          "Entities - Contribution"
        ],
        "summary": "List Contribution records",
        "operationId": "list_Contribution",
        "description": "Retrieve a list of Contribution records with optional filtering, sorting, and pagination.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON query filter, e.g. {\"status\":\"active\"}"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip (pagination)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Field name to sort by. Prefix with '-' for descending order, e.g. -created_date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contribution"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities - Contribution"
        ],
        "summary": "Create a Contribution record",
        "operationId": "create_Contribution",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contribution"
              },
              "example": {
                "word": "Example word",
                "translation_fr": "Example translation_fr",
                "language_code": "Example language_code"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contribution"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Contribution"
        ],
        "summary": "Delete multiple Contribution records",
        "operationId": "delete_many_Contribution",
        "description": "**Warning:** Passing an empty `{}` query will delete ALL records in this entity.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Query filter to match records to delete"
              },
              "example": {
                "word": "Example word"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deleted": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/Contribution/{Contribution_id}": {
      "get": {
        "tags": [
          "Entities - Contribution"
        ],
        "summary": "Get a Contribution record by ID",
        "operationId": "get_Contribution",
        "parameters": [
          {
            "name": "Contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contribution"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities - Contribution"
        ],
        "summary": "Update a Contribution record",
        "operationId": "update_Contribution",
        "parameters": [
          {
            "name": "Contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contribution"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contribution"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Contribution"
        ],
        "summary": "Delete a Contribution record",
        "operationId": "delete_Contribution",
        "description": "Soft-deletes the record. Use the restore endpoint to undo.",
        "parameters": [
          {
            "name": "Contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "api_key",
        "description": "Your API key. Obtain it from your user profile or via `User.me()` in the app SDK."
      }
    },
    "schemas": {
      "Language": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "name_fr": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "flag_emoji": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "coming_soon",
              "dictionary_only"
            ],
            "default": "active"
          },
          "total_lessons": {
            "type": "number",
            "default": 0
          },
          "writing_system": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "Unique record identifier"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record last update timestamp"
          },
          "created_by_id": {
            "type": "string",
            "description": "ID of the user who created the record"
          }
        },
        "required": [
          "code",
          "name",
          "name_fr"
        ],
        "description": "Language entity"
      },
...
      "Contribution": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string"
          },
          "phonetic": {
            "type": "string"
          },
          "translation_fr": {
            "type": "string"
          },
          "language_code": {
            "type": "string"
          },
          "contributor_name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "context_notes": {
            "type": "string"
          },
          "audio_url": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "default": "pending"
          },
          "id": {
            "type": "string",
            "description": "Unique record identifier"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record last update timestamp"
          },
          "created_by_id": {
            "type": "string",
            "description": "ID of the user who created the record"
          }
        },
        "required": [
          "word",
          "translation_fr",
          "language_code"
        ],
        "description": "Contribution entity"
      },
...
    }
  },
...
}